Pin the supersede toast, reshape its copy, drop the dead anchor util
