+++ src/translations/en.json
@@ -680,7 +680,8 @@
     "section_saved": "Section saved",
     "section_save_error": "Failed to save section",
     "section_deleted": "'{name}' section deleted",
-    "delete_superseded": "Deleted on disk, but your unsaved buffer changed during the delete and may still include the section. Review before saving.",
+    "delete_superseded": "Section deleted on disk",
+    "delete_superseded_detail": "Your unsaved buffer changed during the delete and may still include it. Review before saving.",
     "section_delete_error": "Failed to delete section",
     "yaml_saved": "YAML saved",
     "yaml_save_error": "Failed to save YAML",
