+++ test/pages/device-yaml-updated-basis.test.ts
@@ -6,9 +6,11 @@
  * newer draft) advances only the saved side, so the newer draft is
  * neither clobbered nor silently marked clean (#1476).
  */
-import { describe, expect, it } from "vitest";
+import { beforeEach, describe, expect, it, vi } from "vitest";
 
 import "./_mock-device-children.js";
+
+import toast from "sonner-js";
 
 import { ESPHomePageDevice } from "../../src/pages/device.js";
 
@@ -19,6 +21,10 @@ const updated = (page: ESPHomePageDevice, detail: { yaml: string; basedOn: strin
   internals(page)._onYamlUpdated(new CustomEvent("yaml-updated", { detail }));
 
 describe("yaml-updated supersede check", () => {
+  beforeEach(() => {
+    vi.mocked(toast.info).mockClear();
+  });
+
   it("advances both sides when the write's basis is the live buffer", () => {
     const page = new ESPHomePageDevice();
     internals(page)._yaml = "a:\n";
@@ -28,6 +34,7 @@ describe("yaml-updated supersede check", () => {
 
     expect(internals(page)._yaml).toBe("b:\n");
     expect(internals(page)._savedYaml).toBe("b:\n");
+    expect(toast.info).not.toHaveBeenCalled();
   });
 
   it("advances only the saved side when the buffer moved past the basis", () => {
@@ -42,5 +49,8 @@ describe("yaml-updated supersede check", () => {
     // dirty against the disk truth.
     expect(internals(page)._yaml).toBe("draft:\n");
     expect(internals(page)._savedYaml).toBe("b:\n");
+    // The toast is the whole user-visible mitigation for the
+    // retained-buffer divergence.
+    expect(toast.info).toHaveBeenCalledTimes(1);
   });
 });
