+++ test/util/fire-event.test.ts
@@ -1,5 +1,5 @@
 import { describe, expect, it } from "vitest";
-import { fireEvent, fireFromAnchor } from "../../src/util/fire-event.js";
+import { fireEvent } from "../../src/util/fire-event.js";
 
 describe("fireEvent", () => {
   it("dispatches a bubbling composed CustomEvent with the detail", () => {
@@ -27,43 +27,5 @@ describe("fireEvent", () => {
     fireEvent(target, "plain");
 
     expect(seen!.detail).toBeNull();
-  });
-});
-
-describe("fireFromAnchor", () => {
-  it("dispatches from the host while connected", () => {
-    const host = new EventTarget();
-    const anchor = new EventTarget();
-    const seen: string[] = [];
-    host.addEventListener("yaml-updated", () => seen.push("host"));
-    anchor.addEventListener("yaml-updated", () => seen.push("anchor"));
-
-    fireFromAnchor(host, true, anchor, "yaml-updated", { yaml: "x" });
-
-    expect(seen).toEqual(["host"]);
-  });
-
-  it("dispatches from the anchor once the host is disconnected", () => {
-    const host = new EventTarget();
-    const anchor = new EventTarget();
-    const seen: string[] = [];
-    host.addEventListener("yaml-updated", () => seen.push("host"));
-    anchor.addEventListener("yaml-updated", () => seen.push("anchor"));
-
-    fireFromAnchor(host, false, anchor, "yaml-updated", { yaml: "x" });
-
-    expect(seen).toEqual(["anchor"]);
-  });
-
-  it("silently no-ops when disconnected with no anchor", () => {
-    const host = new EventTarget();
-    const seen: string[] = [];
-    host.addEventListener("yaml-updated", () => seen.push("host"));
-
-    // Deliberate contract: a host that never mounted has nowhere to
-    // deliver — swallow rather than throw.
-    fireFromAnchor(host, false, null, "yaml-updated", { yaml: "x" });
-
-    expect(seen).toEqual([]);
   });
 });
